\name{degenes-package}
\alias{degenes-package}
\alias{degenes}
\docType{package}
\title{
Detection of differentially expressed genes.
}
\description{
Detects differentially expressed genes between two distinct groups of samples.
}
\details{
\tabular{ll}{
Package: \tab degenes\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-27\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Klaus Jung

Maintainer: Who to complain to <Klaus.Jung@ams.med.uni-goettingen.de>
}
\references{
Jung K., Quast K., Gannoun A. and Urfer W. (2006) A renewed approach to the nonparametric analysis of replicated microarray experiments. Biometrical Journal, 48, 245-254.
}